fix: better Meetecho API "order" management

+++ ietf/utils/meetecho.py
@@ -480,6 +480,8 @@ class SlidesManager(Manager):
     avoid this requirement.) 
     """
     def add(self, session: "Session", slides: "Document", order: int):
+        # Would like to confirm that session.presentations includes the slides Document, but we can't
+        # (same problem regarding unsaved Documents discussed in the docstring)
         self.api.add_slide_deck(
             wg_token=self.wg_token(session.group),
             session=str(session.pk),
@@ -493,12 +495,36 @@ class SlidesManager(Manager):
         )
 
     def delete(self, session: "Session", slides: "Document"):
+        """Delete a slide deck from the session"""
+        if session.presentations.filter(document=slides).exists():
+            # "order" problems are very likely to result if we delete slides that are actually still
+            # linked to the session
+            raise MeetechoAPIError(
+                f"Slides {slides.pk} are still linked to session {session.pk}."
+            )
+        # remove, leaving a hole
         self.api.delete_slide_deck(
             wg_token=self.wg_token(session.group),
             session=str(session.pk),
             id=slides.pk,
         )
+        if session.presentations.filter(document__type_id="slides").exists():
+            self.send_update(session)  # adjust order to fill in the hole        
     
+    def revise(self, session: "Session", slides: "Document"):
+        """Replace existing deck with its current state"""
+        sp = session.presentations.filter(document=slides).first()
+        if sp is None:
+            raise MeetechoAPIError(f"Slides {slides.pk} not in session {session.pk}")
+        order = sp.order
+        # remove, leaving a hole in the order on Meetecho's side
+        self.api.delete_slide_deck(
+            wg_token=self.wg_token(session.group),
+            session=str(session.pk),
+            id=slides.pk,
+        )
+        self.add(session, slides, order)  # fill in the hole
+        
     def send_update(self, session: "Session"):
         self.api.update_slide_decks(
             wg_token=self.wg_token(session.group),
